+++ packages/at_chops/example/pq/x25519_openssl_dart_ffi.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 import 'dart:ffi';
 import 'dart:typed_data';
 
-import 'package:at_chops/at_chops.dart';
+import 'package:at_chops/at_chops_ffi.dart';
 
 Future<void> main() async {
   final DynamicLibrary? lib = tryLoadLibCrypto();